moved resized mask
