+++ autoarray/mask/mask_2d.py
@@ -602,29 +602,6 @@ class Mask2D(Mask):
             mask=self, sub_size=1, pixel_scales=self.pixel_scales, origin=self.origin
         )
 
-    def resized_mask_from(self, new_shape, pad_value: int = 0.0) -> "Mask2D":
-        """
-        Resized the array to a new shape and at a new origin.
-
-        Parameters
-        ----------
-        new_shape
-            The new two-dimensional shape of the array.
-        """
-
-        mask = copy.deepcopy(self)
-
-        resized_mask = array_2d_util.resized_array_2d_from(
-            array_2d=mask, resized_shape=new_shape, pad_value=pad_value
-        ).astype("bool")
-
-        return Mask2D(
-            mask=resized_mask,
-            pixel_scales=self.pixel_scales,
-            sub_size=self.sub_size,
-            origin=self.origin,
-        )
-
     def trimmed_array_from(self, padded_array, image_shape) -> Array2D:
         """
         Map a padded 1D array of values to its original 2D array, trimming all edge values.
